Move relevant hooks to common package.
