+++ src/hooks/useEpidemicModelState.tsx
@@ -1,9 +1,9 @@
 import { useEffect, useState } from "react";
 
-import { EpidemicModelState } from "../../impact-dashboard/EpidemicModelContext";
-import { LocaleData } from "../../locale-data-context";
-import { Facilities } from "../../page-multi-facility/types";
-import { getEpidemicModelState } from "../responseChartData";
+import { EpidemicModelState } from "../impact-dashboard/EpidemicModelContext";
+import { getEpidemicModelState } from "../infection-model";
+import { LocaleData } from "../locale-data-context";
+import { Facilities } from "../page-multi-facility/types";
 
 export function useEpidemicModelState(
   facilities: Facilities,
