+++ src/page-response-impact/hooks/index.tsx
@@ -1,13 +1,5 @@
-import { useCurrentCurveData } from "./useCurrentCurveData";
-import { useEpidemicModelState } from "./useEpidemicModelState";
 import { useOriginalCurveData } from "./useOriginalCurveData";
 import { usePopulationImpactData } from "./usePopulationImpactData";
 import { useSystemWideData } from "./useSystemWideData";
 
-export {
-  useSystemWideData,
-  useCurrentCurveData,
-  useEpidemicModelState,
-  useOriginalCurveData,
-  usePopulationImpactData,
-};
+export { useSystemWideData, useOriginalCurveData, usePopulationImpactData };
